finished

+++ test/common.js
@@ -7,6 +7,13 @@ const insertUser = body =>
     headers: { 'Content-Type': 'application/json' }
   });
 
+const insertAdminUser = (body, token) =>
+  fetch('http://localhost:8080/admin/users', {
+    method: 'post',
+    body: JSON.stringify(body),
+    headers: { 'Content-Type': 'application/json', Authorization: token }
+  });
+
 const login = body =>
   fetch('http://localhost:8080/users/sessions', {
     method: 'post',
@@ -25,4 +32,4 @@ const getUsers = token =>
     headers: { 'Content-Type': 'application/json', Authorization: token }
   });
 
-module.exports = { insertUser, login, deleteAllUsers, getUsers };
+module.exports = { insertUser, insertAdminUser, login, deleteAllUsers, getUsers };
